[H7][USB] Follow signature changes
USB_LL_Transmit and USB_LL_PrepareReceive:
size was changed from uint16_t to uint32_t

+++ src/main/vcp_hal/usbd_conf_stm32h7xx.c
@@ -594,7 +594,7 @@ USBD_StatusTypeDef USBD_LL_SetUSBAddress(USBD_HandleTypeDef * pdev,
  */
 USBD_StatusTypeDef USBD_LL_Transmit(USBD_HandleTypeDef * pdev,
         uint8_t ep_addr,
-        uint8_t * pbuf, uint16_t size)
+        uint8_t * pbuf, uint32_t size)
 {
     /* Get the packet total length */
     pdev->ep_in[ep_addr & 0x7F].total_length = size;
@@ -613,7 +613,7 @@ USBD_StatusTypeDef USBD_LL_Transmit(USBD_HandleTypeDef * pdev,
  */
 USBD_StatusTypeDef USBD_LL_PrepareReceive(USBD_HandleTypeDef * pdev,
         uint8_t ep_addr,
-        uint8_t * pbuf, uint16_t size)
+        uint8_t * pbuf, uint32_t size)
 {
     HAL_PCD_EP_Receive(pdev->pData, ep_addr, pbuf, size);
     return USBD_OK;
